Revert "more classes moved to platform-gui (#7792)"
This reverts commit 0951605a770d4f97320dd423846106b1614f600d.
