+++ platform-sdk/platform-apps/demos/CryptocurrencyDemo/src/main/java/com/swirlds/demo/crypto/CryptocurrencyDemoMain.java
@@ -142,8 +142,7 @@ public class CryptocurrencyDemoMain implements SwirldMain {
     public void init(final Platform platform, final NodeId id) {
         this.platform = platform;
         this.selfId = id;
-        final int winNum = GuiPlatformAccessor.getInstance().getInstanceNumber(selfId);
-        this.console = createConsole(platform, winNum, true); // create the window, make it visible
+        this.console = createConsole(platform, true); // create the window, make it visible
         GuiPlatformAccessor.getInstance()
                 .setAbout(platform.getSelfId(), "Cryptocurrency and stock market demo v. 1.0\n");
         this.console.addKeyListener(keyListener);
