+++ platform-sdk/platform-apps/demos/HashgraphDemo/src/main/java/com/swirlds/demo/hashgraph/HashgraphDemoMain.java
@@ -410,11 +410,7 @@ public class HashgraphDemoMain implements SwirldMain {
                                 + "Witnesses are colored circles, non-witnesses are black/gray.\n"
                                 + "Dark circles are part of the consensus, light are not.\n"
                                 + "Fame is true for green, false for blue, unknown for red.\n");
-        final Address address = platform.getAddressBook().getAddress(platform.getSelfId());
-        final int winCount = GuiPlatformAccessor.getInstance().getInstanceNumber(platform.getSelfId());
-        final int winNum = GuiPlatformAccessor.getInstance().getInstanceNumber(platform.getSelfId());
-        window = SwirldsGui.createWindow(
-                platform, address, winCount, winNum, false); // Uses BorderLayout. Size is chosen by the Platform
+        window = SwirldsGui.createWindow(platform, false); // Uses BorderLayout. Size is chosen by the Platform
         window.setLayout(new GridBagLayout()); // use a layout more powerful than BorderLayout
         int p = 0; // which parameter to use
         final BiFunction<Integer, String, Checkbox> cb = (n, s) -> new Checkbox(
