+++ platform-sdk/platform-apps/demos/HelloSwirldDemo/src/main/java/com/swirlds/demo/hello/HelloSwirldDemoMain.java
@@ -79,8 +79,7 @@ public class HelloSwirldDemoMain implements SwirldMain {
 
         this.platform = (SwirldsPlatform) platform;
         this.selfId = id;
-        final int winNum = GuiPlatformAccessor.getInstance().getInstanceNumber(selfId);
-        this.console = createConsole(platform, winNum, true); // create the window, make it visible
+        this.console = createConsole(platform, true); // create the window, make it visible
         GuiPlatformAccessor.getInstance().setAbout(platform.getSelfId(), "Hello Swirld v. 1.0\n");
     }
