+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/hashgraph/internal/HashgraphPictureOptions.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2023 Hedera Hashgraph, LLC
+ * Copyright (C) 2022-2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.swirlds.gui.hashgraph;
+package com.swirlds.platform.gui.hashgraph.internal;
 
 /**
- * Options for drawing the {@code HashgraphPicture}
+ * Options for drawing the {@link HashgraphPicture}
  */
 public interface HashgraphPictureOptions {
     /**
