+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/internal/BrowserWindowManager.java
@@ -16,9 +16,9 @@
 
 package com.swirlds.platform.gui.internal;
 
-import com.swirlds.gui.ScrollableJPanel;
 import com.swirlds.platform.SwirldsPlatform;
 import java.awt.GraphicsEnvironment;
+import java.awt.Insets;
 import java.util.ArrayList;
 import java.util.Collection;
 
@@ -33,6 +33,11 @@ public final class BrowserWindowManager {
      * the primary window used by Browser
      */
     private static WinBrowser browserWindow = null;
+
+    /**
+     * the number of pixels between the edges of a window and interior region that can be used
+     */
+    private static Insets insets;
 
     /**
      * metadata about all known apps, swirlds, members, signed states
@@ -59,6 +64,20 @@ public final class BrowserWindowManager {
     }
 
     /**
+     * Get the number of pixels between the edges of a window and interior region that can be used.
+     */
+    public static Insets getInsets() {
+        return insets;
+    }
+
+    /**
+     * Set the number of pixels between the edges of a window and interior region that can be used.
+     */
+    public static void setInsets(final Insets insets) {
+        BrowserWindowManager.insets = insets;
+    }
+
+    /**
      * Get metadata about all known apps, swirlds, members, signed states.
      */
     public static StateHierarchy getStateHierarchy() {
@@ -80,13 +99,14 @@ public final class BrowserWindowManager {
     }
 
     /**
-     * Make the browser window visible. If it doesn't yet exist, then create it. Then switch to the given tab, with a
-     * component name of the form Browser.browserWindow.tab* such as Browser.browserWindow.tabCalls to switch to the
-     * "Calls" tab.
+     * Make the browser window visible. If it doesn't yet exist, then create it. Then switch to the given
+     * tab, with a component name of the form Browser.browserWindow.tab* such as
+     * Browser.browserWindow.tabCalls to switch to the "Calls" tab.
      *
-     * @param comp the index of the tab to select
+     * @param comp
+     * 		the index of the tab to select
      */
-    public static void showBrowserWindow(final ScrollableJPanel comp) {
+    public static void showBrowserWindow(final WinBrowser.ScrollableJPanel comp) {
         showBrowserWindow();
         getBrowserWindow().goTab(comp);
     }
