+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/internal/InfoApp.java
@@ -16,7 +16,6 @@
 
 package com.swirlds.platform.gui.internal;
 
-import com.swirlds.gui.InfoEntity;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -24,13 +23,9 @@ import java.util.List;
  * Metadata about an app that is installed locally.
  */
 public class InfoApp extends InfoEntity {
-    private List<InfoSwirld> swirlds = new ArrayList<>(); // children
+    List<InfoSwirld> swirlds = new ArrayList<InfoSwirld>(); // children
 
     public InfoApp(String name) {
-        super(name);
-    }
-
-    public List<InfoSwirld> getSwirlds() {
-        return swirlds;
+        this.name = name;
     }
 }
