+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/internal/InfoSwirld.java
@@ -16,7 +16,6 @@
 
 package com.swirlds.platform.gui.internal;
 
-import com.swirlds.gui.InfoEntity;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -24,22 +23,15 @@ import java.util.List;
  * Metadata about a swirld running on an app.
  */
 public class InfoSwirld extends InfoEntity {
+    public InfoApp app; // parent
+    List<InfoMember> members = new ArrayList<InfoMember>(); // children
 
-    private final InfoApp app; // parent
-
-    private final List<InfoMember> members = new ArrayList<>(); // children
+    Reference swirldId;
 
     public InfoSwirld(InfoApp app, byte[] swirldIdBytes) {
-        super("Swirld " + new Reference(swirldIdBytes).to62Prefix());
         this.app = app;
-        this.app.getSwirlds().add(this);
-    }
-
-    public InfoApp getApp() {
-        return app;
-    }
-
-    public List<InfoMember> getMembers() {
-        return members;
+        this.swirldId = new Reference(swirldIdBytes);
+        name = "Swirld " + swirldId.to62Prefix();
+        app.swirlds.add(this);
     }
 }
