+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/internal/WinBrowser.java
@@ -16,18 +16,19 @@
 
 package com.swirlds.platform.gui.internal;
 
+import static com.swirlds.common.system.SystemExitUtils.exitSystem;
 import static com.swirlds.logging.LogMarker.EXCEPTION;
 import static com.swirlds.platform.gui.internal.BrowserWindowManager.getBrowserWindow;
 import static com.swirlds.platform.gui.internal.BrowserWindowManager.showBrowserWindow;
 
-import com.swirlds.gui.GuiConstants;
-import com.swirlds.gui.GuiUtils;
-import com.swirlds.gui.ScrollableJPanel;
-import com.swirlds.gui.hashgraph.HashgraphGuiSource;
+import com.swirlds.common.system.SystemExitCode;
+import com.swirlds.gui.PrePaintableJPanel;
+import com.swirlds.platform.gui.hashgraph.HashgraphGuiSource;
 import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.Component;
 import java.awt.Dimension;
+import java.awt.Font;
 import java.awt.Graphics;
 import java.awt.GraphicsEnvironment;
 import java.awt.GridBagLayout;
@@ -37,8 +38,11 @@ import java.awt.event.ActionListener;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import java.awt.event.MouseListener;
+import java.awt.event.WindowAdapter;
+import java.awt.event.WindowEvent;
 import javax.swing.JFrame;
 import javax.swing.JPanel;
+import javax.swing.JScrollPane;
 import javax.swing.JTabbedPane;
 import javax.swing.JTextArea;
 import javax.swing.Timer;
@@ -66,6 +70,8 @@ public class WinBrowser extends JFrame {
 
     /** refresh the screen every this many milliseconds */
     final int refreshPeriod = 500;
+    /** use this font for all text in the browser window */
+    static Font FONT = new Font("SansSerif", Font.PLAIN, 16);
     /** the InfoMember that is currently being shown by all tabs in the browser window */
     static volatile InfoMember memberDisplayed = null;
     /** have all the tabs been initialized yet? */
@@ -75,6 +81,8 @@ public class WinBrowser extends JFrame {
     private static Timer updater = null;
     /** gap at top of the screen (to let you click on app windows), in pixels */
     private static final int topGap = 40;
+    /** light blue used to highlight which member all the tabs are currently displaying */
+    static final Color MEMBER_HIGHLIGHT_COLOR = new Color(0.8f, 0.9f, 1.0f);
 
     static ScrollableJPanel tabSwirlds;
     static ScrollableJPanel tabAddresses;
@@ -93,6 +101,45 @@ public class WinBrowser extends JFrame {
     public void paintComponents(Graphics g) {
         super.paintComponents(g);
         Thread.currentThread().setPriority(Thread.MAX_PRIORITY);
+    }
+
+    /**
+     * A JScrollPane containing the given UpdatableJPanel.
+     *
+     * @author Leemon
+     */
+    public static class ScrollableJPanel extends JScrollPane {
+        private static final long serialVersionUID = 1L;
+        PrePaintableJPanel contents;
+
+        /**
+         * Wrap the given panel in scroll bars, and remember it so that calls to prePaint() can be passed on
+         * to it.
+         */
+        ScrollableJPanel(PrePaintableJPanel contents) {
+            super(contents);
+            this.contents = contents;
+            contents.setVisible(true);
+        }
+
+        /**
+         * Recalculate the contents of each Component, maybe slowly, so that the next repaint() will trigger
+         * a fast render of everything.
+         */
+        void prePaint() {
+            if (contents != null) {
+                contents.prePaint();
+            }
+        }
+    }
+
+    /** set the component to have a white background, and wrap it in scroll bars */
+    private ScrollableJPanel makeScrollableJPanel(PrePaintableJPanel comp) {
+        comp.setBackground(Color.WHITE);
+        ScrollableJPanel scroll = new ScrollableJPanel(comp);
+        scroll.setBackground(Color.WHITE);
+        scroll.setVisible(true);
+        return scroll;
     }
 
     /** when clicked on, switch to the Swirlds tab in the browser window */
@@ -116,10 +163,10 @@ public class WinBrowser extends JFrame {
 
             // perform the equivalent of prePaint on nameBarName:
             if (WinBrowser.memberDisplayed != null) {
-                nameBarName.setText("    " + WinBrowser.memberDisplayed.getName() + "    ");
+                nameBarName.setText("    " + WinBrowser.memberDisplayed.name + "    ");
             } else { // retry once a second until at least one member exists. Then choose the first one.
                 if (tabSwirlds != null) {
-                    ((WinTabSwirlds) tabSwirlds.getContents()).chooseMemberDisplayed();
+                    ((WinTabSwirlds) tabSwirlds.contents).chooseMemberDisplayed();
                 }
             }
             // call prePaint() on the current tab, only if it has such a method
@@ -199,12 +246,12 @@ public class WinBrowser extends JFrame {
         nameBarLabel = new JTextArea();
         nameBarName = new JTextArea();
         tabbed = new JTabbedPane(JTabbedPane.TOP, JTabbedPane.WRAP_TAB_LAYOUT);
-        tabSwirlds = GuiUtils.makeScrollableJPanel(new WinTabSwirlds());
-        tabAddresses = GuiUtils.makeScrollableJPanel(new WinTabAddresses());
-        tabCalls = GuiUtils.makeScrollableJPanel(new WinTabCalls());
-        tabPosts = GuiUtils.makeScrollableJPanel(new WinTabPosts());
-        tabNetwork = GuiUtils.makeScrollableJPanel(new WinTabNetwork(hashgraphSource));
-        tabSecurity = GuiUtils.makeScrollableJPanel(new WinTabSecurity());
+        tabSwirlds = makeScrollableJPanel(new WinTabSwirlds());
+        tabAddresses = makeScrollableJPanel(new WinTabAddresses());
+        tabCalls = makeScrollableJPanel(new WinTabCalls());
+        tabPosts = makeScrollableJPanel(new WinTabPosts());
+        tabNetwork = makeScrollableJPanel(new WinTabNetwork(hashgraphSource));
+        tabSecurity = makeScrollableJPanel(new WinTabSecurity());
 
         Rectangle winRect = GraphicsEnvironment.getLocalGraphicsEnvironment().getMaximumWindowBounds();
 
@@ -216,7 +263,7 @@ public class WinBrowser extends JFrame {
         setLocation(winRect.x, winRect.y + topGap);
         setFocusable(true);
         requestFocusInWindow();
-        addWindowListener(GuiUtils.stopper());
+        addWindowListener(stopper());
 
         nameBar.setLayout(new GridBagLayout());
         nameBar.setPreferredSize(new Dimension(1000, 42));
@@ -250,7 +297,7 @@ public class WinBrowser extends JFrame {
         setBackground(Color.WHITE); // this color flashes briefly at startup, then is hidden
         nameBar.setBackground(Color.WHITE); // color of name bar outside label and name
         nameBarLabel.setBackground(Color.WHITE); // color of name bar label
-        nameBarName.setBackground(GuiConstants.MEMBER_HIGHLIGHT_COLOR); // color of name
+        nameBarName.setBackground(MEMBER_HIGHLIGHT_COLOR); // color of name
         tabbed.setBackground(Color.WHITE); // color of non-highlighted tab buttons
         tabbed.setForeground(Color.BLACK); // color of words on the tab buttons
 
@@ -276,5 +323,27 @@ public class WinBrowser extends JFrame {
         requestFocus(true);
         tabbed.setSelectedComponent(contents);
         prePaintThenRepaint();
+    }
+
+    /**
+     * Add this to a window as a listener so that when the window closes, so does the entire program,
+     * including the browser and all platforms.
+     *
+     * @return a listener that responds to the window closing
+     */
+    static WindowAdapter stopper() {
+        return new WindowAdapter() {
+            public void windowClosed(WindowEvent e) {
+                exitSystem(SystemExitCode.NO_ERROR, "window closed", true);
+            }
+        };
+    }
+
+    /**
+     * Instantiates and returns a JTextArea whose settings are suitable for use inside the browser window's
+     * scroll area in a tab.
+     */
+    static JTextArea newJTextArea() {
+        return GuiUtils.newJTextArea("");
     }
 }
