+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/internal/WinTab2Consensus.java
@@ -16,9 +16,8 @@
 
 package com.swirlds.platform.gui.internal;
 
-import static com.swirlds.gui.GuiUtils.wrap;
+import static com.swirlds.platform.gui.internal.GuiUtils.wrap;
 
-import com.swirlds.gui.GuiUtils;
 import com.swirlds.gui.PrePaintableJPanel;
 import com.swirlds.platform.Consensus;
 import com.swirlds.platform.SwirldsPlatform;
@@ -44,7 +43,7 @@ class WinTab2Consensus extends PrePaintableJPanel {
      * Instantiate and initialize content of this tab.
      */
     public WinTab2Consensus() {
-        text = GuiUtils.newJTextArea("");
+        text = WinBrowser.newJTextArea();
         add(text);
     }
 
@@ -55,7 +54,7 @@ class WinTab2Consensus extends PrePaintableJPanel {
             if (WinBrowser.memberDisplayed == null) {
                 return;
             }
-            SwirldsPlatform platform = WinBrowser.memberDisplayed.getPlatform();
+            SwirldsPlatform platform = WinBrowser.memberDisplayed.platform;
             String s = "";
             s += GuiPlatformAccessor.getInstance().getPlatformName(platform.getSelfId());
             final Consensus consensus = GuiPlatformAccessor.getInstance().getConsensus(platform.getSelfId());
