+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/internal/WinTab2Stats.java
@@ -17,15 +17,13 @@
 package com.swirlds.platform.gui.internal;
 
 import static com.swirlds.common.metrics.Metric.ValueType.VALUE;
-import static com.swirlds.gui.GuiUtils.wrap;
 import static com.swirlds.logging.LogMarker.EXCEPTION;
+import static com.swirlds.platform.gui.internal.GuiUtils.wrap;
 
 import com.swirlds.common.config.BasicConfig;
 import com.swirlds.common.metrics.LegacyMetric;
 import com.swirlds.common.metrics.Metric;
 import com.swirlds.common.metrics.statistics.internal.StatsBuffer;
-import com.swirlds.gui.GuiConstants;
-import com.swirlds.gui.GuiUtils;
 import com.swirlds.gui.PrePaintableJPanel;
 import java.awt.Color;
 import java.awt.Dimension;
@@ -89,7 +87,7 @@ class WinTab2Stats extends PrePaintableJPanel {
      * Instantiate and initialize content of this tab.
      */
     public WinTab2Stats() {
-        metrics = WinBrowser.memberDisplayed.getPlatform().getContext().getMetrics().getAll().stream()
+        metrics = WinBrowser.memberDisplayed.platform.getContext().getMetrics().getAll().stream()
                 .sorted(Comparator.comparing(m -> m.getName().toUpperCase()))
                 .toList();
         int numStats = metrics.size();
@@ -337,7 +335,7 @@ class WinTab2Stats extends PrePaintableJPanel {
 
                 if (buffer.numBins() == 0) {
                     final BasicConfig basicConfig = WinBrowser.memberDisplayed
-                            .getPlatform()
+                            .platform
                             .getContext()
                             .getConfiguration()
                             .getConfigData(BasicConfig.class);
@@ -439,7 +437,7 @@ class WinTab2Stats extends PrePaintableJPanel {
         int j = cm2rm[index];
         chartsPanel.add(charts[index]);
         charts[index].setVisible(true); // flag this as having been added
-        statBoxes[j].setBackground(GuiConstants.MEMBER_HIGHLIGHT_COLOR);
+        statBoxes[j].setBackground(WinBrowser.MEMBER_HIGHLIGHT_COLOR);
         revalidate();
     }
 
@@ -477,8 +475,8 @@ class WinTab2Stats extends PrePaintableJPanel {
         chartsPanel.setBackground(Color.WHITE);
         setBackground(Color.WHITE);
 
-        JTextArea instructions = GuiUtils.newJTextArea("");
-        JTextArea spacer = GuiUtils.newJTextArea("");
+        JTextArea instructions = WinBrowser.newJTextArea();
+        JTextArea spacer = WinBrowser.newJTextArea();
         instructions.setText(wrap(
                 50,
                 ""
@@ -513,7 +511,7 @@ class WinTab2Stats extends PrePaintableJPanel {
             c.weightx = 0;
             boxesPanel.add(txt, c);
             if (i % numCols == numCols - 1) {
-                JTextArea spacer2 = GuiUtils.newJTextArea("");
+                JTextArea spacer2 = WinBrowser.newJTextArea();
                 c.gridx++;
                 c.weightx = 1.0f;
                 c.gridwidth = GridBagConstraints.REMAINDER; // end of row
@@ -548,7 +546,7 @@ class WinTab2Stats extends PrePaintableJPanel {
                 showChart(j);
             }
         }
-        descriptions = GuiUtils.newJTextArea("");
+        descriptions = WinBrowser.newJTextArea();
         descriptions.setFont(new Font(Font.MONOSPACED, Font.PLAIN, 14));
         final StringBuilder s = new StringBuilder();
         for (final Metric metric : metrics) {
@@ -574,7 +572,7 @@ class WinTab2Stats extends PrePaintableJPanel {
         add(chartsPanel, c);
         c.gridy = 3;
         add(descriptions, c);
-        JTextArea spacer2 = GuiUtils.newJTextArea("");
+        JTextArea spacer2 = WinBrowser.newJTextArea();
         c.gridy = 4;
         c.weighty = 1.0f;
         add(spacer2, c);
